requirements.txt
Django>=4.2.4
psycopg2-binary>=2.9.6
clickhouse-driver>=0.2.7
gunicorn>=21.1.0
requests[socks]>=2.32.3
urllib3==1.26.15
django-modeltranslation>=0.18.6
django-admin-list-filter-dropdown>=1.0.3
uvicorn[standard]>=0.30.1
boto3>=1.34.132
python-dateutil>=2.9.0.post0
django-unfold
yadisk==3.4.0